# lib/api-zod/src/generated/types/createReminderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediNova API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReminderBodyFrequency } from "./createReminderBodyFrequency";

export interface CreateReminderBody {
  medicationName: string;
  dosage: string;
  frequency: CreateReminderBodyFrequency;
  times: string[];
  startDate: Date;
  endDate?: Date | null;
  notes?: string | null;
  color?: string | null;
}