# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediNova API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalReminders: number;
  activeReminders: number;
  todayTotal: number;
  todayTaken: number;
  todayMissed: number;
  /** Percentage 0-100 */
  adherenceRate: number;
  /** Consecutive days of full adherence */
  streak: number;
}