# lib/api-zod/src/generated/types/logDoseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediNova API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogDoseBodyStatus } from "./logDoseBodyStatus";

export interface LogDoseBody {
  status: LogDoseBodyStatus;
  takenAt?: Date | null;
  scheduledTime: string;
}