# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediNova API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderFrequency } from "./reminderFrequency";

export interface Reminder {
  id: number;
  medicationName: string;
  dosage: string;
  frequency: ReminderFrequency;
  /** Array of HH:MM strings */
  times: string[];
  startDate: Date;
  endDate?: Date | null;
  notes?: string | null;
  isActive: boolean;
  color?: string | null;
  createdAt: Date;
}