# lib/api-zod/src/generated/types/todayReminder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediNova API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TodayReminderStatus } from "./todayReminderStatus";

export interface TodayReminder {
  id: number;
  medicationName: string;
  dosage: string;
  scheduledTime: string;
  status: TodayReminderStatus;
  color?: string | null;
}